improved structure, added javadoc

+++ src/main/java/com/helvetica/Model/DeviceComparator.java
@@ -4,8 +4,14 @@ import java.util.Comparator;
 
 public class DeviceComparator implements Comparator<Device> {
 
+    /**
+     * Overridden Comparator method
+     * @param comparableDevice (Device)
+     * @param deviceToCompare (Device)
+     * @return (int) to compare by
+     */
     @Override
     public int compare(Device comparableDevice, Device deviceToCompare) {
-        return (int) (comparableDevice.getPower() - deviceToCompare.getPower());
+        return comparableDevice.getPower() - deviceToCompare.getPower();
     }
 }
